Fix slow ssh/ping when in test VMs
This was due to the fact that without assigning the bridge interface a
static MAC address. It would use the same one on both VMs.

Pretty sure this means any packet sent with that MAC address would go to
both VMs. Meaning the behavior probably isn't going to be what I
expected.