Rearrange yaml to be more aware of state
This is to ensure changes occur only when needed (e.g. gpg key changes).

As for the network configuration, if the bridge exists, then we can
assume that the machine has been configured. At least, this is based on
when ansible_default_ipv4.interface is the bridge interface.

While using ansible_default_ipv4.interface is not perfect, its a reasonable
implementation for this.