Add initial new caic role, refactor leftover junk
The container creation and initial configuration role has been added to
handle the details of the above. There was an attempt to use ctrservers as
the 'hosts' for the play, but that ran into initial issues with the
community.general.lxd_container module. Even still, once the containers
were created, Ansible would assume it would connect to them via ssh.
This was not desired.

The role iterates over each its tasklist for each container, with the
top play providing the ctrserver to connect to and create the container.
Thus, another iteration. Clustering is currently OOS.

The development inventory file has its ctrserver group now refactored to
mimic how my production inventory uses alias for its host(s).