Setup ctrservers w/containers and python install
Most of the changes made, or the implementation derive from the
reddit post mentioned in the 'Create LXD container nodes on ctrservers'
play.

caic role turns out not to be needed but that implementation was only
able to ever create a container but never able to connect to it (at least
with a local connection).