Add initial Vagrantfile and tweaks
Using 10.10.100.x as the test subnet turned out to be a bad idea as
lxc-source configured the server's bridge to be in that subnet as well.
Hence a new default subnet has been added as part of the LXD preseed.

The vagrant-customizations playbook exists to add configuration that I
don't wish for the site playbook to worry about (e.g. boot parameters).