Refactorings, change uri to get_url
The uri module (builtin) module is meant more for interacting with
web services but can possibly download just and html document.
get_url I believe is more appropriate for the use cases I so far in the
project.

The systemd_networkd tasks/handlers have received the become attribute
because whether or not different permissions are needed should be
at the discretion of the role itself and the not the play the role is
included in.

The pkg parameter for apt (builtin) has been changed to name, as I'd
rather not use an alias (it appears these can be deprecated).