Change custom lineinfile implementation
The local lineinfile implementation now adds a Ansible managed header
initially if one does not exist (and if the line does not already exist
in some form in the file).

The builtin lineinfile does do replacing based on the regexp passed in.
That said, if the regexp matches both a expected line initially and the
line after being replaced, then idempotence will be achieved. Meaning,
having the new line replaced again should not report a change back to
Ansible.

Whereas, I've had to play around and remember that this seems to only
be an issue when the line to be inserted contains newlines (in which
case a check should be perform prior for the contents aside the newline
(s)).

lbyl prefix has been changed with managed to reflect that lines to be
inserted will now fall under a Ansible managed header.