Separate ctrservers into top level playbook
This is to match Ansible's example of how to organize playbook content
or at least one of the ways they (the organization) recommend.

Update cache/index for apt has been moved to the top as a separate task.
This allows other calls to apt to not have to worry about performing an
update on the cache.

If the grammer expressed is just talking about 'LXD' (e.g. not LXD
service, LXD container, ...)
then 'the' should not be prepended.