Move lxd_source password to a playbook
The lxd_source password has been moved to be sourced during a playbook
from a variable file. I wish to use Ansible Vault to encrypt secrets so
storing these in a file I felt was best.

The encrypted secrets are stored in Bitwarden and pulled when needed.
It would not be a great idea to push the encrypted
secrets and have it be tracked by git.

Small ansible-lint refactors were done as well. If I felt a shell
command was too long, then I use a literal block scalar instead of
fitting everything on one line.