Eliminate majority of unnecessary quoting
This was done mainly as things were somewhat inconsistent across the
yaml configurations. In general, it seems not necessary to quote
strings.

That said, quotes were left/changed if needed like in the
lxc-containers.yaml. limits.cpu needs to be quoted otherwise Ansible
threw an error when trying to append the configuration to the existing
instance configuration.