Reorder systemd-networkd and install lxd plays
Previously the plays order in the playbook didn't flow the best in
my opinion.

That said, additional patching had to be added as of recent when
learning Docker's appendage of rules to the iptables configuration. For
more information, see the comment under the
"Insert iptables rules for compatibility between LXD and Docker" play.