Rework iptables rules for LXD and Docker
It was found that by using changing the default iptables utility to the
legacy tool(s)(iptables and friends) that LXD would insert its own
iptables rules. Thus, it does not appear that I have todo this manually.

A bug has been fixed where ansible_env['HOME'] would actually point to
root's home instead of the actual ansible user's home. This caused the
actual python binary to not be found. The fix was to manually set the
ansible user's home.