Add version parameters for items, move add_host
Including the jenkins-torkel docker container, jenkins-infrastructure
repo. This allows from a high-level for these values to be modified in
one place.

In addition for jenkins-infrastructure, if this repo does update, then
the project configuration file is removed and regenerated. This is
mainly incase any specific changes to the configuration file occur.

Container hosts will now be added as inventory based on tasks from a
different file. This mainly because I do want to have all inventory
variables being added for each container.