Add ability for sequential installs of LXD
Prior to this change, I only assumed that a new version of LXD could be
dropped in place. A few additional changes were needed to make this
happen.

Due to the LD_LIBRARY_PATH being changed when running Ansible, the steps
needed to ensure sequential installs needed to be split in two.