Add TODO, continue using global python3-apt libs
While it is possible to build out the python(3)-apt libs and insert them
into virtualenv site-packages directory, there isn't a striking reason
todo this.

The only usage I've seen of those libs are when Ansible uses the
builtin apt task module. Which, isn't enough of a reason to have a
separate version in the virtualenv, at least for now.