Tag ctrserver plays, and group some plays together
It was noticed in conjunction with the Python virtualenv, that trying to
group blocks of tasks under this virtualenv proved to be somewhat
difficult and would have only deteriorated the flow of tasks.

Instead, I opted to just group the initial setup tasks together and leave
LXD/Docker installations and the rest as separate plays.

ansible.compatibility_mode is needed as an unknown version of Ansible
causes vagrant to assume defaults to the older 1.8 version of
Ansible. Which is not desired, and I have no need to use a version of
Ansible that old.