Check over absolute path usage throughout configs
A majority of the copy tasks copied over literal content over to a new
file (e.g. not copying a file from A to B), which in case needs an
absolute path.

GitHub directories perhaps could change and not use a "repo_name"
variable, however, that means the repo will be cloned wherever
Ansible happens to be sitting at in the filesystem on the client system.
Which is not something I want.

Only change would be for the extracted LXD archive directory, but mainly
as the prefix for the extracted was assumed for 2 tasks (and by
extension the task in charge of extracting the archive). I wouldn't
expect the prefix to change but its better to be safe.