Refactor project based on ansible-lint 5.3.1
According to
https://github.com/ansible-community/ansible-lint/issues/1398#issuecomment-782879721,
starting with ansible-lint 5.0.0, the linter now expects the project
to be structured in a specific way to indentify yaml files in terms of a
Ansible context (e.g. playbook vs tasklist).

This structure can be interpreted from the following link:
https://github.com/ansible-community/ansible-lint/blob/main/src/ansiblelint/config.py#L13-L33.

Aside that, other changes to Ansible files were based on the linter's
complaints.