Implement changes based on gps Jenkins job testing
A make wrapper has now been created for Jenkins projects to use. This is
due to my emphasis on having either the actual project or infrastructure
worry about certain environment needs for Jenkins projects. This is in
concerns currently because of pyenv shell shenanigans needed and myself
not wanting each project to have to source a file for each shell step.

The pyenv root directory should be entirely owned by the jenkins user.
Otherwise I noticed strange errors due to the jenkins user not having
sufficient permissions for some files in the pyenv root directory.

Finally, each altaria container needs to connect to a Jenkins agent as
the jenkins user. Otherwise each build of a Jenkins job/project will run
as root. Which is probably not a great idea and not what I believe I
wanted.