Provision irc_client locally
For now and where possible, I will spin up VPS(s) locally to test. This
could change drastically if my VPS(s) in the cloud are to talk to my
machines locally. This is because network configurations will change
heavily between the physical homelab and the hypervisor based homelab.

A slight caveat is that the cloud environment will not entirely mimic a
local virtual setup. However, for quick testing purposes, I think this
is ok.