Add initial implementation to diagram homelab
Making use of mingrammer's diagrams software, this is in attempt to make
strides in automating the documentation of my homelab infrastructure.
Other tools like drawio and lucidcharts require more manual labor to
keep their diagrams update to date but I did at least consider them.
Having had used lucidcharts in the past.

Eventually I would like to store network configurations in a common
place and have the diagram script refer to them (1+ because it is
written in python). However, I see this being the next iterative step
in this process.