Use jenkins_base_containers role vs hand rolling
The setup and instantiation of my jenkins-base containers has been
consolidated into a role for somewhat easier maintenance. Hence instead
of explicity defining the task(s) and parameter(s) for my jenkins-base
containers, this will be done implicity through a role.

The jenkins-torkel version has been incremented as I've delete the
previous version from Docker Hub and I really did not want to have to
hand role that image again (the Makefile did not have a rule to create
the Docker image as of that version).