Discard particular docker & container.io pkgs
In contrast to 6a20c885, the goal going forward is to try and lessen the
burden of managing the upgrading of software versions. While the idea of
having this fine granular control seems neat on paper, the time cost
would probably be more vs doing incremental upgrades and hoping for the
best.

I certainly do not have time to review each and every change made to a
package. Hence I'd probably just upgrade the packages to the latest,
test, and hope for the best anyways.