Replace jenkins-infrastructure with jenkins role
This was done mainly due to two things.

One, cloning the jenkins-infrastructure and doing all the shenanigans
needed to run a script that would connect a machine to Jenkins does not
seem like the best approach to tackle this problem. Instead, this
functionality and others related have been moved to a jenkins role.
This puts more emphasize on Ansible.

Two, to continue using the jenkins-infrastructure repo, I would have to
package this repo to be distributed for my homelab. An approach that I
feel is more work than its worth.