Improve disunion btwn lightsail/irc_client configs
The prior lightsail role had configurations that were not specific to
lightsail itself but rather to configuring a machine to be an
irc_client. Hence, a majority of the tasks that lightsail had has been
moved to the irc_client's playbook.

For the moment, any vagrant customizations to make up the differences
between a machine in the cloud vs locally has been removed. While some
differences can be worked around, I'm ok knowing that a local VM will
not perfectly replicate a configured machine in the cloud.

Variable names/locations has been adjusted to be more clear on intent.

To adjust for the notion that an irc_client might run on different
platforms (e.g. aws, gcp, locally), the lightsail role will only be
invoked if that particular irc_client is also part of the
lightsail_instances group. The production inventory file has been
changed to reflect this by using further subgrouping.