Make privilege escalation consistent (altaria)
The assumption should be that unless the ansible user needs to run a
task as a privileged user, then it shouldn't. I noticed in the altaria
machine configurations, this wasn't consistent due to the primary usage
of these configurations on LXD containers. Which by default with the LXD
connection plugin, uses root as the remote_user/ansible_user.

The jenkins_user as of current, or this writing, does not need privilege
escalation for operations it may need to perform. However, in the event
this changes, I should look into a more specific sudoers configuration.