Remove reqs for docker and general collections
At least for the community namespace. This is because those two
collections are already part of the "ansible" package and as of this
commit, there isn't an exact version needed for either of these
collections.

That said, by locking the version of ansible used, this effectively
should lock the version of these collections until the ansible package
version is updated.