Integrate the bitwarden common mk file
I originally wanted to put the setting of SKIP_BITWARDEN_GET_SSH_KEYS
and SKIP_BITWARDEN_GET_TLS_CERTS after the common Makefiles are
included, however, setting these variables after caused the behavior to
be incorrect (e.g. ssh keys were still grabbed). Hence, for some reason,
these variables need to be set before the common Makefiles are imported.

If in the future the ansiscrts target has behavior that does not
explicity use the bitwarden CLI tool, then the appropriate skip variable
will need to be set. This ensures at least for that behavior, the
bitwarden CLI session is not checked. It might even be possible that the
session check no longer needs to exist as a prerequisite.