Prevent duplicate entries in libvirt network xml
It was noticed that the resulting network configuration for the homelab
management libvirt network had duplicate host elements for at least each
ctrserver. It appears that the VM's provider configuration loops more
than once (in fact, when using 'puts "#{machine_name}"' inside of the VM
provider configuration, their were 15 entries in stdout) and I have a
bet its due to the lazy loading that occurs inside of multi-machine
configurations.

The solution in place is more of a work around than it is a solution to
the lazy loading that occurs inside of multi-machine configurations.
What is important is that any future implementations working inside of
multi-machine configurations need to ensure their actions are
idempotent. Meaning, additional runs do affect the end result.

Even though technically it does not seem to have made a difference that
duplicate static host IP entries existed in the network configuration
up to this point I had no issues), it is something I wanted to address.