Integrate a dhcp node into the infrastructure
Turns out the setup with vagrant needed more changes than I initially
thought. For starters, each spun up VM will now have 2 nics each. One
nic will be on the vagrant management network and will be solely for
allowing vagrant to connect and provision a VM. The second nic will be
the mock project network to use that should simulate what will actually
exist on production.

The management network still needs to be hand crafted so I can more
deterministically know when that network is created and add static
entries to it's DHCP server. Without static entries, when the machines
restart (if one uses the systemd-networkd role), Ansible will be unable
to reconnect to the VMs.

Due to lazy loading, (related to 5e6aa767) I also noticed that not all
VMs get a static entry in the management network's DHCP server. All but
the last VM will get an entry. As of this commit, irc-client1 will not
get a static entry but this is ok because it does not restart it's
networking daemon (systemd-networkd) or have other conditions that could
possibly land it with a different IP address. Just a footnote going
forward that this works out for now but could change in the future where
a dummy VM might need to be created. Even uncondensing the inner
configurations (e.g. manually writing out each VM config) does not seem
to fix this.