Make dhcp_listen_ipv4_addr a host var
If the variable in the subject line remained in the
'network_configs.yml' file, then any other DHCP servers would also that
exact IP address as their listening address. While not an issue at the
moment, I figured making the variable host based wouldn't be hard as I
was already doing this for the listening interface's MAC address.

A implementation has been added in the Vagrantfile to handle a JSON of
configurations (in the machine's host variables) whose values will be
evaluated to those in the parent JSON. This was done to prevent
duplication of values and allow dhcp_listen_ipv4_addr to be a host
variable.