Add initial implementation for dnsmasq's dns
homelab_network_subnet in VAGRANT_HOMELAB_NETWORK_CONFIGS should have
been set originally to VAGRANT_LIBVIRT_HOMELAB_NETWORK_SUBNET and not
the management variant.

The dnsmasq_dns role contains a default template with generic variables
that are meant to be specific at a "higher" level (e.g. config file or
playbook).

The DNS template to be used for production currently is not finished and
will be revisited once I have a better idea on the resources that will
need to be pointed after clustering numerous software pieces together.