Move 'Restart dnsmasq' to a local handlers role
While it appears collections' roles can use each others handlers without
additonal importing, this does not seem to be the case with projects'
roles. Hence, to allow roles to use common handlers, the play importing
the local role needs to also import the local common handlers.

While inconsistent at the role level, there is consistency at the play
level in that you are importing the common handlers whether they are
local or not.