Pass in DHCP/DNS default vars explicitly to roles
This would be consistent with how other tasks assign values to the
variables in a role. In that, there is some high level variable that
eventually is assigned to the applicable variable in a role.

This was technically already done but the assignment of variables is
more explicit by being done right when calling both the DHCP and DNS
roles.