Revert 8d4d649 because vars unused in custom tpl
While passing in variables to a role explicitly would keep things
consistent with the rest of my Ansible code, it doesn't make sense here
todo so. Because more than likly a custom template will be used to
define both dnsmasq's DHCP and DNS configuration, the role should not
care if the default variables are passed in or not. Not passing them in
should help signify that the default template is not going to be used
from the role.

Even the template to be used should be a inventory host variable and not
a hard coded value in the playbook. As I do expect, at least mostly for
DHCP, to use a different template if a different DHCP server is needed
for some reason. This means each template to each DHCP and DNS server
should be unique, thus generic variables are in templates are not
required for now (the vagrant dns template still contained these).