Integrate the Calico CNI
Helm was chosen to install Calico because it appears that the version of
Calico is tied to its Helm chart. This was because its manifest did not
appear to contain any notation of its software version.
The version of Calico may eventually be tied in from somewhere.

The worker and controller (staging) nodes will now use the
generic/ubuntu2204 Vagrant box instead of
debian/bullseye64 due to an issue where the CNI loopback plugin could
not be found in /usr/lib/cni. While this solution does not directly
address the issue, it does get around said issue and for now, I am ok
with this that the containerd version is more recent and not as likely
to cause conflicts with the newer Kubernetes version).

The encapsulation setting for Calico currently is 'None'. As this
appears to be the most common configuration for on-prem clusters with
Calico and I do not currently have a need to have an overlay setup.