Add additional k8s controllers serially to cluster
I noticed when doing my Kubernetes cluster setup, that it would be
common to see at least once that either controller 2 or 3 would fail to
join the cluster as an additional control plane. The error present would
say something like, "etcdserver: re-configuration failed due to not
enough started members".

It appears joining additional control planes concurrently is somewhat
flaky in whether or not it would work. For now, my solution to this
would be to join the additional control planes serially vs concurrently.
For reference on a related GitHub issue with the same error seen:
https://github.com/kubernetes/kubeadm/issues/2001