Remove setting static network settings on machines
Due to running a DHCP server (at least on the homelab subnet), I find
setting static network settings for those machines that already have
dedicated static IPs set on the DHCP server to be redundant and
mutually exclusive (e.g. an interface either sets a static IP or listens
to DHCP for an IP).

Also, in the event I ever change a machine's homelab IP, I have to not
only change this on the DHCP server (in the event a static IP is set)
but I also have to update the machine's static network settings for the
new IP (assuming it doesn't set the interface settings to just listen on DHCP).

The exception to the above as of this writing should only be the ansible-controller (which currently also serves as the homelab's DHCP and DNS server).