Add migration to nftables research for k8s
In short, while there exists some workarounds to getting a Kubernetes
cluster (regardless of variant) to work with a nftables backend, I do
not think it is currently worth it to integrate such into my homelab
Kubernetes cluster.

This would require additional configurations and testing that would be
an another burden that eventually would need to be change when
Kubernetes finally supports nftables natively.