Integrate jenkins-torkel into Kubernetes cluster
While ideally, I would prefer to have more than one replica for the
jenkins-torkel controller pod, the gain from having more than one
replica vs the work needed I don't feel is enough currently to justify
having more than one replica.

While it might be useful to save JENKINS_HOME to some form of persistent
storage like an NFS server, again, there isn't much currently to gain
from all of that work. Maybe some build logs could be saved, but that's
all I think that could be worth keeping. That, and I've had issues
before with jobs not updating because JENKINS_HOME is not overwritten
with the new jobs listing. However, that could be related to the casc
file placement in the image.

I'm choosing to leave JENKINS_HOME to be volume mounted in the
jenkins-torkel container as a footnote. Even though currently it is not
mounted with anything useful.