Run ingress ngnix tasks post controller init
The ingress ngnix Helm chart deploys a Kubernetes job in which creates a
pod on the cluster. However, when creating the cluster from scratch, the
pod fails to be created due to only the first controller being the only
node running in the cluster at the time the tasks run. That, and control
planes are not typically scheduled with normal pods on them.

While another solution would just to run these tasks on the worker nodes
, by default (even with sudo) helm on these nodes does not have the
right permissions to modify the cluster according to the chart.

Hence, the solution was to run another playbook after the entire cluster
has been initialized to install the ingress ngnix helm chart.