Genericize k8s configs
It was between a solution like this or separating the (poseidon)
Kubernetes configurations into a separate directory of its own at the
top level of the repository. Mainly Because of the upfront work required
todo the latter, and to some lesser extent the potential of copypasta
between the Poseidon cluster and any future new Kubernetes cluster.
Despite making use of roles/task lists where I can.

In the ansibile_groups, I have refactored the groups to be generic first
(e.g. group of groups -> group of instances), the same goes with the
respective variables section inside of ansible_groups.

While it is possible that I may want to use a different deployment tool
in the future to deploy Kubernetes (e.g. Kubespray), a subtle limitation
of this refactoring is that both the k8s_control_plane and
k8s_worker_node roles use kubeadm and currently do not take in
consideration of the deployment tool used.

Calico isn't directly a pre_task to deploy Jenkins, but I would prefer
to have network shenanigans out the way prior to deploying any
Kubernetes services.

During the refactoring, I ran into where the Kubernetes setup would
complain about /proc/sys/net/bridge/bridge-nf-call-iptable not existing
and I have run into this before. That said, it wasn't clear to me how
fix this at the time but it turns out the br_netfilter module needs to
be loaded first before adding the respective kernel parameters.