Reduce future number of templates needed
Going forward, templates that are not intended for a specific
environment will not contain an environment prefix (e.g. `dev`, `prod`,
`staging`). Previously, templates created to be used by the
vagrant/staging environment would have a `vagrant-` prefix appended to
their template name. Templates meant for `production` would not have any
prefix appended.

Have examining all of the templates, I determined that the majority of
them should work in any environment. Hence, the majority of templates
did not receive an environment prefix.

The default template name used for the haproxy role had to be changed
because due to local relative path searching by Ansible. If the default
template has the same name as the playbook template, the default
template would be picked up before the playbook template. Which is not
the behavior I desire. The default template name now resembles what the
roles dnsmasq_dns and dnsmasq_dhcp use.