Generalize dnsmasq_dns, dnsmasq_dhcp, and haproxy
These changes were inspired by commit cd303b96.

The destination filename for the evaluated jinja template is now exposed
as a variable for the roles mentioned in the subject. This is to make
the haproxy role like the roles dnsmasq_dns and dnsmasq_dhcp but now
with all exposing the destination filename as a variable.