Pair keepalived with haproxy
The integration of keepalived was inspired by the 'High Availability
Considerations' documentation from kubeadm.

For the health check script, instead of performing a request to the load
balancer, I opted to just ensure the load balancer was running. This is
because otherwise when keepalived starts and runs the script, none of
the kube-apiservers will be running. This causes the script to not exit
and keeps any virtual IP from being assigned. This might be something I
return to, however.