Add remaining NICs for ansible-controller1
When using systemd-networkd, it appears that any NICs not explicitly
mentioned are not picked up automatically by systemd-networkd. This
isn't true in some scenarios like with the Kubernetes nodes but that is
only because the nodes' Linux distro (Ubuntu) uses netplan. netplan has
configurations that will write NIC device configurations to
systemd-networkd's volatile runtime network directory.