Add ability for config_ref to refer to a ruby var
The syntax of the config_ref to refer to a ruby runtime variable will be
that of a shell-like variable. While using eval can be "nasty", the
situation where this is used in the Vagrantfile draws little concern.

config_refs don't nesscessarily always represent a machine attribute,
despite the comment mentioned in eval_config_ref.