Discard notion of a ansible controller
Reverts 0235c64. I've decided now that it is not worth it to integrate a
ansible controller at this time. Instead, a generic node will be named
that will serve as the DHCP/DNS/Load Balancer node. These
functionalities will be given a separate playbook each to keep match the
existing pattern of playbooks.

The http_frontend IP address variable name has changed in a way that the
template can remain environment generic. (e.g. probably won't have a
staging-node1 machine in production).