Manage bitwarden's CLI client via npm
This will make it easier to keep the CLI version up to date for the
project. Though I do imagine this will require re-testing the applicable
targets that use the CLI client. I am ok with at the moment.

Modern versions of npx are just part of npm now. Hence, no need to check
if the npx binary can be found in the PATH.