Refactor Ruby dependencies' installations
The XML library (nokogiri) that I did install through 'make setup' has
been removed because Vagrant's embedded Ruby interpreter already has
this library installed, probably never using the one I installed.
e2689e0 does not detail why I felt like it was needed to install this
library outside of Vagrant.

Bundle is needed to install gems via Gemfile. It's about how pip,
despite being a third package to Python, is pretty much always shipped
with Python. That relationship is how I would describe Bundle and Ruby.
The reason this was done this way is to allow the version of the plugin
to be managed by Renovate. Vagrant does not have a file format that
Renovate could manage for its plugins (at least at the time of writing
this).