Check for default grub config on affected asks
If the (initial) check for the boot parameter in the default grub
configuration file fails because the default grub configuration file
does not exist, line_check will not be defined correctly, affecting
other tasks. This should be countered to an extent by using the
default_grub_config_inode.stat.exists expression in a when conditional
for a task.

Most subsequent tasks do use a when conditional with the expression
mentioned above, but a few don't. There doesn't appear to be a reason
for this either looking back at 28895b1b and I would argue that it was a
mistake not to use a when conditional with said expression.