Reposition where vagrant_config_refs is deleted
Previously I noticed that the line where vagrant_config_refs is removed
from the machine_attrs was not indented correctly. However, upon looking
further, I also wondered why this line was at the bottom of the "each"
block.

There doesn't appear to be any prior reason for this though deleting a
non-existing hash doesn't cause Ruby to throw an exception. To prevent
further confusion, I will move this into the block that knows
vagrant_config_refs exists as a attribute of machine_attrs.

Also, there was a block that was missing a closing 'end' to it. That has
also been fixed.