Fix staging-node1's eth2 interface getting ip addr
As of this writing, previously, I assumed that the mentioned interface
was set with a static IP address. This however was not the case when
looking at the systemd-networkd device files, this has been corrected by
assigning the interface with a static IP address and not getting it via
DHCP.

It's possible during caf6136d, that I forgot no DHCP server is actually
active on the network that the interface serves.