Remove setting static IP for nodes in test network
Considering the original intent of this test (net-) network was to test
DHCP and DNS, it doesn't make sense for any of the machines but the DHCP
node itself to have a static IP address set. It might provide usefully
though to continue setting the MAC address for other machines too, so
for now, that will stay.