Convert some host vars to env generic group vars
The granularity provided by making these variables host variables is
unnecessary because, for example, all Poseidon Kubernetes controllers
(first and high availability) install and use containerd. Unless new
plays are written to install a different container runtime and target a
specific controller, which would then warrant the cri_socket_path
variable to be a host variable (which could just exist for that
particular controller and not all of them), the current implementation
would be better off having this variable be an environment generic group
variable in a group variable file in the project.