Migrate weechat role into irc_clients playbook
It is at this point in the project where I have felt the need to slim
down the number of internal roles that exist. Since this role is not
used more than once internally and has a unlikely chance of being needed
externally, I have decided to migrate the weechat role into the
irc_clients playbook.