Utilize devices' SSH key pair more heavily
Instead of using the default regional key pair or a separate key pair to
access Milo from my machines, I've opted to add the public keys of my
machines to Milo's authorized_keys for the 'admin' and 'weechat' users.
It is expected my machines will have the correct identity file in the
default path (e.g. ~/.ssh/id_rsa).

The pulling of the private key that can access my GitHub account I
believe is a remnant of f65805f8. There is no reason as of now to
download that key for local use.