Add the Metrics Server to the Poseidon k8s cluster
serverTLSBootstrap needs to be set to true to allow secure connections
(TLS) between the metrics-server and the nodes' kubelet.

address and node-ip are set to the variable kubelet_listen_ipv4_addr
because I currently do not see a situation where the node's kubelet
listens on an IP address different from the node's cluster IP address.
Originally the node's IP address registered to the cluster would be the
default IP address of the said node. This should now be set to its
cluster IP address.