Add active health checks to haproxy instances
By default, HAProxy will try to 'verify' the server's certificate when
setting up an SSL connection with the respective servers that an HTTP
check is being performed on. Hence, the 'ca-file' must exist on the
filesystem prior to running HAProxy. To ensure this, all Poseidon
controllers now receive their root CA certificate prior to the cluster
being initialized.

The first controller still uploads the Kubernetes certificates to be
distributed to the other controllers. The other controllers still
retrieve these certificates, but this overwrites the previously copied
root CA certificate. If that occurs, I wanted to make sure HAProxy was
loaded with the same CA certificate that is written on the filesystem.