Install etcdctl onto the Poseidon k8s controllers
This was added to help in the event an etcd quorum was lost.

While it seems etcdctl is released with the same version of etcd, I
don't see any restrictions regarding the version of etcdctl that must be
used with a version of etcd. The version of etcdctl was picked based on
what kubeadm uses for etcd. Hence if the version of ectd is upgraded,
then I intend on upgrading the version of etcdctl to that version as
well.