Install cri-tools as part of the k8s_kubeadm role
It was noticed that cri-tools 1.26.00-00 causes a kudeadm pre-flight
test to fail with the following error: "validate service connection: CRI
v1 runtime API is not implemented for endpoint
unix:///run/containerd/containerd.sock". Essentially, it seems many
Kubernetes components are moving to deprecate using older CRI versions
and instead use the stable v1 interface going forward. That, and it
appears containerd 1.5 support also ends with Kubernetes v1.26.

To ensure cluster functionality, for now, an older version of cri-tools
will need to be installed.