Upgrade Poseidon Kubernetes version to v1.25
The template kubeletconfiguration.yaml.j2 does break the convention on
how I name my YAML jinja templates but according to the Kubernetes
documentation (v.1.25), the patch extension must either be 'json' or
'yaml'.

kubeadm, kubectl, and kubelet package versions needed to be one patch
lower from the Kubernetes version to be compatible with the package the
version of cri-tools.