Integrate a vmm node that runs QEMU/KVM & libvirt
To keep consistent with production, I've added a Poseidon Kubernetes
network to staging and have appended the appropriate nodes to it (not
all nodes need to be connected to this network). In a different commit I
will create the production variant of the Poseidon libvirt-network.xml
template.

The vrrp_interface has changed for all the Poseidon Kubernetes
controllers because ultimately keepalived allows for further enhancement
to the high availability setup of kube-apiservers. Hence, this should be
listening on the same interface used for my Poseidon Kubernetes traffic.

mgmt_ipv4_addrs have been shifted back down 1 host address because
homelab_poseidon_vrrp_server_vip no longer exists on the mgmt-homelab-cm
network.

Inspiration for the vmms.yml playbook has come from the
dhcp_servers.yml, dns_servers.yml, and load_balancers.yml playbooks. In
that, the default virtual machine monitor (vmm) to be installed will be
QEMU/KVM.