Refactor Ansible tags
Each tag should be prefixed with a verb consistent with other tags.
Currently, unless it is Kubernetes related, other playbooks don't need
the 'initial' prefix added to them or their tags. This is because the
Kubernetes playbooks currently warrant the additional prefix.

The on-premises setup play name has been refactored to be consistent
with other playbooks.