Add k8s-node packer images
The Ubuntu autoinstall config does not parameterize the ansible_user
because at the moment, it is not intended currently that I will actually
provision the nodes on the staging virtual machine monitor aside from
uploading the virtual disks and creating the VMs from that.

The 'cloud-init status --wait' was added because I want to ensure that
the cloud-init initialization processes finish before exporting the
image. This part was recommended in a Packer AMI tutorial by Canonical.

The artifacts from the builds are renamed, so I have control of the
output name in the proper location (vs dealing with this in an Ansible
playbook).