Update GPG key checksums
The k8s_kubeadm role now processes the Kubernetes's (Google's) GPG key
by dearmoring the file before installing it onto the system. This was
done because before this, I was getting an error much like
https://github.com/kubernetes/release/issues/2862. This solution follows
the recommendations from the PR that derived from this issue.

Dearmoring was not done originally because Kubernetes v1.24
documentation (at the time and as of now) for installing kubeadm does
not instruct this when downloading the key.